requirements.txt
Django>=4.1,<4.2
wagtail>=4.2,<4.3

djangorestframework~=3.14.0